% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSE_wrap.R
\name{EHR_MuSe}
\alias{EHR_MuSe}
\title{IPW and AIPW Methods for Multi-cohort Selection Bias in Non-probability Samples}
\usage{
EHR_MuSe(
  K,
  Z_names,
  intdata_list,
  N = NULL,
  UW_CS = FALSE,
  IPW = FALSE,
  weights_user = NULL,
  AIPW = FALSE,
  ipw_method = "PL",
  extdata = NULL,
  marginals_list = NULL,
  select_var_list = NULL,
  aux_var_list = NULL,
  Weights_e = NULL,
  aux_model = "XGBoost",
  variance = FALSE,
  type_var = "approx"
)
}
\arguments{
\item{K}{\itemize{
\item Necessary Input. Number of cohorts. Should be a numeric positive integer.
}}

\item{Z_names}{\itemize{
\item Necessary Input. A character vector containing the names of the Z variables or disease model covariates.
}}

\item{intdata_list}{\itemize{
\item Necessary Input. A list of size K where each element of list corresponds to the data for each of the K multiple cohorts including the disease indicator D, the Z variables and the selection variables in the disease model. Each data should be of the form of a data frame. Please include a column named "id" to indicate unique identifiers to the units.
}}

\item{N}{\itemize{
\item Target Population Size.
}}

\item{UW_CS}{\itemize{
\item An indicator variable (TRUE or FALSE) for using the unweighted logistic regression model with cohort-specific intercepts.
}}

\item{IPW}{\itemize{
\item An indicator variable (TRUE or FALSE) for using the Inverse Probability Weighted Methods (IPW) methods.
}}

\item{weights_user}{\itemize{
\item User specified weights. A numeric vector of weights for the combined data (not duplicated).
}}

\item{AIPW}{\itemize{
\item An indicator variable (TRUE or FALSE) for using the Joint Augmented IPW method. If AIPW is TRUE, please also input IPW to be TRUE.
}}

\item{ipw_method}{\itemize{
\item If IPW is TRUE, specify the IPW method to be used. A character variable. Default is PL (Pseudolikelihood). Other options are Simplex Regression (SR), Calibration (CL) or User Specified (US).
}}

\item{extdata}{\itemize{
\item If IPW method is set to PL or SR or AIPW is TRUE, please provide a data frame containing individual level external data which is a probability sample like NHANES. The external data should contain all the selection variables and if AIPW is TRUE, then also all the auxiliary score model variables.  Please include a column named "id" to indicate unique identifiers to the units.
}}

\item{marginals_list}{\itemize{
\item If IPW method is set to CL, please provide a list of size K in which each element is a numeric vector containing the marginal sums of the selection variables of each of the K cohorts. Please ensure the first element for each of the K numeric vector should be the population size, N.
}}

\item{select_var_list}{\itemize{
\item If IPW is set to be TRUE, please provide a list of size K in which each element is a character vector corresponding to the selection variables' names for each of the K cohorts.
}}

\item{aux_var_list}{\itemize{
\item If AIPW is set to be TRUE, please provide a list of size K in which each element is a character vector corresponding to the auxiliary score model variables' names for each of the K cohorts.
}}

\item{Weights_e}{\itemize{
\item If IPW method is set to PL or SR or AIPW is TRUE, please provide the known selection weights for the external probability sample. The input should be a numeric vector.
}}

\item{aux_model}{\itemize{
\item If AIPW is true, please provide the auxiliary score model. Default is XGboost.
}}

\item{variance}{\itemize{
\item An indicator variable (TRUE or FALSE) whether variance should be computed or not, along with the point estimate.
}}

\item{type_var}{\itemize{
\item If variance is true, indicate the method type to be used for computing the variance. For the unweighted method, do not provide any type. For IPW methods, PL and CL, we have two options, asy (asymptotic variance incorporating the variance from nuisance parameters) and "approx" ignoring the variance from nuisance parameters. For SR and AIPW, we have only the approx method. The default for IPW and AIPW methods is approx.
}}
}
\description{
IPW and AIPW Methods for Multi-cohort Selection Bias in Non-probability Samples
}
\examples{
res_uw=MUSE(K=K,N=N,Z_names=Z_names,intdata_list=intdata_list,variance = TRUE)

}
